casting to keep compiler happy

+++ src/simplexy/src/dallpeaks.c
@@ -30,8 +30,8 @@ static int *yc = NULL;
 int objects_compare(const void *first, const void *second)
 {
 	float v1, v2;
-	v1 = dobject[*((int *) first)];
-	v2 = dobject[*((int *) second)];
+	v1 = (float)dobject[*((int *) first)];
+	v2 = (float)dobject[*((int *) second)];
 	if (v1 > v2)
 		return (1);
 	if (v1 < v2)
@@ -41,6 +41,7 @@ int objects_compare(const void *first, const void *second)
 
 /* Finds all peaks in the image by cutting a bounding box out around each one
  * */
+//FIXME -- sam says why is there a parameter maxnpeaks?
 
 int dallpeaks(float *image,
               int nx,
@@ -69,7 +70,7 @@ int dallpeaks(float *image,
 			dobject[i + j*nx] = objects[i + j * nx];
 	for (i = 0;i < nx*ny;i++)
 		indx[i] = i;
-	qsort((void *) indx, nx*ny, sizeof(int), objects_compare);
+	qsort((void *) indx, (size_t) nx*ny, sizeof(int), objects_compare);
 	for (l = 0;l < nx*ny && dobject[indx[l]] == -1; l++)
 		;
 	nobj = 0;
